+++ deployment/terraform/iam.tf
@@ -79,6 +79,27 @@ data "aws_iam_policy_document" "anonymous_read_storage_bucket_policy" {
   }
 }
 
+data "aws_iam_policy_document" "anonymous_read_tile_cache_bucket_policy" {
+  policy_id = "S3TileCacheAnonymousReadPolicy"
+
+  statement {
+    sid = "S3ReadOnly"
+
+    effect = "Allow"
+
+    principals {
+      type        = "AWS"
+      identifiers = ["*"]
+    }
+
+    actions = ["s3:GetObject"]
+
+    resources = [
+      "arn:aws:s3:::${lower(var.environment)}-pfb-tile-cache-${var.aws_region}/*",
+    ]
+  }
+}
+
 #
 # Custom policies
 #
@@ -145,8 +166,8 @@ resource "aws_iam_role_policy_attachment" "ecs_for_ec2_policy_container_instance
 }
 
 resource "aws_iam_instance_profile" "app_container_instance" {
-  name  = "${aws_iam_role.app_container_instance_ec2.name}"
-  role =  "${aws_iam_role.app_container_instance_ec2.name}"
+  name = "${aws_iam_role.app_container_instance_ec2.name}"
+  role = "${aws_iam_role.app_container_instance_ec2.name}"
 }
 
 #
@@ -168,6 +189,6 @@ resource "aws_iam_role_policy_attachment" "batch_ec2_s3_policy" {
 }
 
 resource "aws_iam_instance_profile" "batch_container_instance" {
-  name  = "${aws_iam_role.batch_container_instance_ec2.name}"
+  name = "${aws_iam_role.batch_container_instance_ec2.name}"
   role = "${aws_iam_role.batch_container_instance_ec2.name}"
 }