+++ src/angularjs/src/app/neighborhoods/create/neighborhoods-create.controller.js
@@ -10,13 +10,21 @@
     'use strict';
 
     /** @ngInject */
-    function NeighborhoodCreateController($state, $filter, toastr, Upload, Neighborhood, State) {
+    function NeighborhoodCreateController($log, $state, $filter, toastr, Upload, Neighborhood,
+                                          Country, State) {
         var ctl = this;
 
+        var DEFAULT_COUNTRY = {abbr: 'US', name: 'United States'};
+
         function initialize() {
+            ctl.country = DEFAULT_COUNTRY;
             ctl.states = State.query();
             ctl.states.$promise.then(function(response) {
                 ctl.states = response;
+            });
+            ctl.countries = Country.query();
+            ctl.countries.$promise.then(function(response) {
+                ctl.countries = response;
             });
             // TODO: De-dupe from API?
             ctl.visibilities = [
@@ -36,7 +44,9 @@
                 method: 'POST',
                 data: {
                     boundary_file: ctl.file,
-                    state_abbrev: ctl.state.abbr,
+                    state_abbrev: ctl.state && ctl.isDefaultCountry() ? ctl.state.abbr : '',
+                    city_fips: ctl.city_fips && ctl.isDefaultCountry() ? ctl.city_fips : '',
+                    country: ctl.country.abbr || DEFAULT_COUNTRY,
                     visibility: ctl.visibility,
                     label: ctl.label
                 }
@@ -45,10 +55,31 @@
                 toastr.success('Successfully created neighborhood');
                 $state.go('admin.neighborhoods.list');
             }).catch(function(error) {
+                $log.error(error);
                 toastr.clear(uploadToast);
-                toastr.error('Unable to create neighborhood: ' + error);
+                var msg = 'Unable to create neighborhood:';
+                if (error.data && error.data.non_field_errors) {
+                    // extract non-field errors
+                    for (var i in error.data.non_field_errors) {
+                        msg += ' ' + error.data.non_field_errors[i];
+                    }
+                } else if (error.data) {
+                    // extract field errors
+                    for (var err in error.data) {
+                        msg += ' ' + err + ': ' + error.data[err];
+                    }
+                    $log.error(msg);
+                } else {
+                    msg += ' ' + error;
+                }
+                toastr.error(msg);
+                $log.error(msg);
             });
         };
+
+        ctl.isDefaultCountry = function() {
+            return ctl.country && ctl.country.abbr === 'US';
+        }
     }
     angular
         .module('pfb.neighborhoods.create')