+++ src/tilegarden/src/api.js
@@ -3,9 +3,11 @@
  */
 
 const APIBuilder = require('claudia-api-builder')
+const aws = require('aws-sdk')
 
 const { imageTile, createMap } = require('./tiler')
 const HTTPError = require('./util/error-builder')
+const logger = require('./util/logger')
 
 const IMAGE_HEADERS = {
     'Content-Type': 'image/png',
@@ -15,12 +17,13 @@ const HTML_RESPONSE = { success: { contentType: 'text/html' } }
 
 // Converts a req object to a set of coordinates
 const processCoords = (req) => {
+    logger.debug('processCoords')
     // Handle url params
-    const z = Number(req.pathParams.z)
-    const x = Number(req.pathParams.x)
+    const z = Number(req.pathParameters.z)
+    const x = Number(req.pathParameters.x)
 
     // strip .png off of y if necessary
-    const preY = req.pathParams.y
+    const preY = req.pathParameters.y
     const y = Number(preY.substr(0, preY.lastIndexOf('.')) || preY)
 
     // Check type of coords
@@ -32,31 +35,20 @@ const processCoords = (req) => {
 }
 
 const getPositionalFilters = (req) => {
+    logger.debug('getPositionalFilters')
     /* eslint-disable-next-line object-curly-newline */
-    const { x, y, z, config, ...remainder } = req.pathParams
+    const { x, y, z, config, ...remainder } = req.pathParameters
     return remainder
-}
-
-// Returns a properly formatted list of layers
-// or an empty list if there are none
-const processLayers = (req) => {
-    if (req.queryString.layers) return JSON.parse(req.queryString.layers)
-    else if (req.queryString.layer || req.queryString.filter || req.queryString.filters) {
-        /* eslint-disable-next-line quotes */
-        throw HTTPError("Invalid argument, did you mean '&layers='?", 400)
-    }
-
-    return []
 }
 
 // Parses out the configuration specifications
 const processConfig = req => ({
-    s3bucket: req.queryString.s3bucket,
-    config: req.pathParams.config,
+    s3bucket: req.queryStringParameters ? req.queryStringParameters.s3bucket : null,
+    config: req.pathParameters.config,
 })
 
 // Create new lambda API
-const api = new APIBuilder()
+const api = new APIBuilder({ requestFormat: 'AWS_PROXY' })
 
 // Handles error by returning an API response object
 const handleError = (e) => {
@@ -69,6 +61,48 @@ const handleError = (e) => {
     )
 }
 
+/* Uploads a tile to the S3 cache, using its request path as a key
+ *
+ * Does nothing unless there's a CACHE_BUCKET set in the environment.
+ * Returns the Promise<tile> again for chaining.
+ *
+ * Theoretically it should be possible to run the upload in parallel and not make the request
+ * wait for it before returning the tile, but in fact the process gets killed when the main promise
+ * resolves so the upload doesn't manage to finish.
+ */
+const writeToS3 = (tile, req) => {
+    logger.debug('writeToS3 called')
+    const s3CacheBucket = process.env.PFB_TILEGARDEN_CACHE_BUCKET
+    if (s3CacheBucket) {
+        let key = req.path
+        // API Gateway includes a 'path' property but claudia-local-api currently doesn't
+        // (see https://github.com/azavea/claudia-local-api/issues/1), so this reconstructs it.
+        if (!key) {
+            logger.debug('writeToS3: recreating key from path params')
+            /* eslint-disable camelcase */
+            const { z, x, y, job_id, config } = req.pathParameters
+            key = `tile/${job_id}/${config}/${z}/${x}/${y}`
+            /* eslint-enable camelcase */
+        } else if (key.startsWith('/')) {
+            // API Gateway request.path object starts with a leading slash,
+            // which would cause the uploaded object to have the wrong prefix.
+            key = key.slice(1)
+        }
+
+        logger.debug('writeToS3: uploading')
+        const upload = new aws.S3().putObject({
+            Bucket: s3CacheBucket,
+            Key: key,
+            Body: tile,
+        })
+        return upload.promise().then(() => {
+            logger.info(`Uploaded tile to S3: ${key}`)
+            return tile
+        })
+    }
+    return new Promise(resolve => resolve(tile))
+}
+
 // Get tile for some zxy bounds
 api.get(
     '/tile/{job_id}/{config}/{z}/{x}/{y}',
@@ -76,10 +110,11 @@ api.get(
         try {
             const { z, x, y } = processCoords(req)
             const filters = getPositionalFilters(req)
-            const layers = processLayers(req)
             const configOptions = processConfig(req)
 
-            return imageTile(createMap(z, x, y, filters, layers, configOptions))
+            logger.debug('api.get: creating imageTile')
+            return imageTile(createMap(z, x, y, filters, configOptions))
+                .then(tile => writeToS3(tile, req))
                 .then(img => new APIBuilder.ApiResponse(img, IMAGE_HEADERS, 200))
                 .catch(handleError)
         } catch (e) {