+++ src/tilegarden/tests/api.test.js
@@ -1,12 +1,11 @@
 const rewire = require('rewire')
 const api = rewire('../src/api'),
-    processCoords = api.__get__('processCoords'),
-    processLayers = api.__get__('processLayers')
+    processCoords = api.__get__('processCoords')
 
 describe('processCoords', () => {
     test('properly parses properly formatted coords', () => {
         const req = {
-            pathParams: {
+            pathParameters: {
                 x: '2385',
                 y: '3103.png',
                 z: '13'
@@ -21,14 +20,14 @@ describe('processCoords', () => {
 
     test('parses extensions other than png', () => {
         const req = {
-            pathParams: {
+            pathParameters: {
                 x: '2385',
                 y: '3103.jpg',
                 z: '13'
             }
         }
         const req2 = {
-            pathParams: {
+            pathParameters: {
                 x: '2385',
                 y: '3103.pdf',
                 z: '13'
@@ -48,7 +47,7 @@ describe('processCoords', () => {
 
     test('properly parses coords with no file extension', () => {
         const req = {
-            pathParams: {
+            pathParameters: {
                 x: '2385',
                 y: '3103',
                 z: '13'
@@ -63,7 +62,7 @@ describe('processCoords', () => {
 
     test('throws an error for non-numeral coords', () => {
         const req = {
-            pathParams: {
+            pathParameters: {
                 x: 'eggs',
                 y: 'bacon',
                 z: 'orange_juice'
@@ -73,38 +72,3 @@ describe('processCoords', () => {
     })
 })
 
-describe('processLayers', () => {
-    test('properly parses list of layer', () => {
-        const req = {
-            queryString: {
-                layers: '["layer1","layer2","layer3","layer4"]'
-            }
-        }
-        expect(processLayers(req)).toEqual(['layer1','layer2','layer3','layer4'])
-    })
-
-    test('properly parses just one layer', () => {
-        const req = {
-            queryString: {
-                layers: '["layer"]'
-            }
-        }
-        expect(processLayers(req)).toEqual(['layer'])
-    })
-
-    test('properly parses no fields', () => {
-        const req = {
-            queryString: {}
-        }
-        expect(processLayers(req)).toEqual([])
-    })
-
-    test('properly parses a blank field', () => {
-        const req = {
-            queryString: {
-                layers: ''
-            }
-        }
-        expect(processLayers(req)).toEqual([])
-    })
-})